refactor pm quality fairness detail router
